+++ src/private/initialize/Get-KeepAChangelogHeadingLineList.ps1
@@ -1,4 +1,4 @@
-function Get-KeepAChangelogHeadingLines {
+function Get-KeepAChangelogHeadingLineList {
     [CmdletBinding()]
     param(
         [Parameter(Mandatory)]